Prevent main page loading too often
Main page is loading twice, once `onCreate` and `onResume`, with the
latter overriding whatever happens on creation. The used field
`lastLanguage` is never written, so `onResume` will always load the main
page resulting in application resets when the is reactivated.

As the language change never resets the application, doing such a check
(with a working field) is not necessary and the language reset (and main
page loading) can be done when actually changing the language. The
initial loading on application creation is one in `onCreate`. As such
overriding `onResume` is not necessary.