Refactor file URL loading
Load file information simultaneously with the normal title loading. For
non-File titles, no additional data is shown. All File titles are within
namespace 6 and can be identified unambiguously without checking for all
possible localizations of the File namespace prefix. In case of
namespace 6 (a File title), image information is included in the API
response and the data can be used to start the image display instead.